Calculate connected components specifications (#1)

+++ build.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 clear
 printf '\e[3J'
-rm -rf build
+if [ -d "build" ]; then
+    rm -rf build
+fi
 target=$1
 cmake -S . -B build
 cmake --build build --config Debug --target $target
+if [ ! -d "out" ]; then
+  mkdir out
+fi
 ./build/$target
